Update README.md
Inventory management refers to the process by which you track the amount of product you have on your warehouse shelf, in store or sitting with other retailers and distributors. This enables you to succeed in having the right number of units in the right place, at the right time and for the right price.